đa luồng tìm kiếm và xem userprofile

+++ src/main/java/app/tiktok/category/ChallengeInfo.java
@@ -45,5 +45,5 @@ public class ChallengeInfo {
 	/** The number of users who have uploaded a video for the challenge */
 	 Integer user_count;
 	 
-	 Integer view_count;
+	 Long view_count;
 }
